requirements.dev.txt
pytest
django-pytest
django-stubs
mypy
